Squash owt 88-sdk

+++ modules/audio_processing/agc2/rnn_vad/features_extraction.cc
@@ -76,12 +76,12 @@ bool FeaturesExtractor::CheckSilenceComputeFeatures(
   // Analyze reference and lagged frames checking if silence has been detected
   // and write the feature vector.
   return spectral_features_extractor_.CheckSilenceComputeFeatures(
-      reference_frame_view_, {lagged_frame.data(), kFrameSize20ms24kHz},
+      reference_frame_view_, {lagged_frame.data(), (size_t)kFrameSize20ms24kHz},
       {feature_vector.data() + kNumLowerBands, kNumBands - kNumLowerBands},
-      {feature_vector.data(), kNumLowerBands},
-      {feature_vector.data() + kNumBands, kNumLowerBands},
-      {feature_vector.data() + kNumBands + kNumLowerBands, kNumLowerBands},
-      {feature_vector.data() + kNumBands + 2 * kNumLowerBands, kNumLowerBands},
+      {feature_vector.data(), (size_t)kNumLowerBands},
+      {feature_vector.data() + kNumBands, (size_t)kNumLowerBands},
+      {feature_vector.data() + kNumBands + kNumLowerBands, (size_t)kNumLowerBands},
+      {feature_vector.data() + kNumBands + 2 * kNumLowerBands, (size_t)kNumLowerBands},
       &feature_vector[kFeatureVectorSize - 1]);
 }
